karışık

+++ code/modules/antagonists/heretic/heretic_living_heart.dm
@@ -55,7 +55,10 @@
 /datum/component/living_heart/proc/on_organ_replaced(obj/item/organ/source, obj/item/organ/replacement)
 	SIGNAL_HANDLER
 
-	if(IS_ROBOTIC_ORGAN(replacement))
+	// PSYCHONAUT EDIT ADDITION BEGIN - IPC - Original:
+	// if(IS_ROBOTIC_ORGAN(replacement))
+	if(IS_ROBOTIC_ORGAN(replacement) && !istype(replacement, /obj/item/organ/heart/cybernetic/tier2/ipc))
+	// PSYCHONAUT EDIT ADDITION END - IPC
 		qdel(src)
 		return
